+++ src/commands/units/refershUnitsExplorer.ts
@@ -15,10 +15,10 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import { IActionContext } from "@dviererbe/vscode-utils";
-import { UnitsTreeItem } from "../../trees/units/UnitsTreeItem";
+import { ExtensionVariables } from "../../extensionVariables";
 
-export function refershUnitsExplorer(context: IActionContext, unitsTree: UnitsTreeItem): Promise<void>
+export function refershUnitsExplorer(context: IActionContext): Promise<void>
 {
-    return unitsTree.refresh(context);
+    return ExtensionVariables.unitsRoot.refresh(context);
 }